requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.1
nltk==3.8.1
scikit-learn==1.5.1
gensim==4.3.3
transformers==4.42.4
torch==2.3.1
joblib==1.4.2
fastapi==0.111.1
uvicorn==0.30.3
pyodbc==5.1.0
sqlalchemy==2.0.31
scipy==1.11.4
google-generativeai==0.7.2